Remove unneeded #memoize, #transform, #transform_unless methods
We previously discovered that #transform and #transform_unless were quite slow
and stopped using them for that reason. While #memoize is potentially useful,
its functionality is tangential to the reason for this library to exist: to provide
efficient, easy-to-use persistent data structures for Ruby.

Users who need something like #memoize can easily subclass one of the immutable
collections and add this functionality. There is no reason why it has to be included
right in "core". It could even be spun out into a separate library.

This cuts down on a little bit of extraneous overhead for users who do not
need #memoize.