Fix tests for Immutable.to_ruby for Ruby 3.0
In Ruby 2, SortedSet was part of the standard library. In Ruby 3, it
was removed from stdlib and put into a gem. This new gem implements
SortedSet using RBTree (from the 'rbtree' gem). As a result, the
behavior of `SortedSet#eql?` has changed; two SortedSets only compare
as equal using `#eql?` if they are the same object, not if their
contents are the same.

This behavior does not match built-in Ruby data structures like `Hash`
and is arguably incorrect. It was reported as an issue in the GitHub
repository for `sorted_set` almost a year ago, but has not been fixed.

We would love to see this issue with the new `sorted_set` gem fixed,
but in the meantime, we can simply avoid using `SortedSet#eql?` for
testing equality in our specs.