Check arity of comparator block for SortedSet correctly for Ruby 3.0
The behavior of `&:method` blocks has changed in Ruby 3.0. Previously
they had arity -1; now it's -2. I don't know the reason for this
change.

While we fix that, we can also be more strict in rejecting blocks with
an arity which we can't use. These would result in an exception
anyways, but it will be better to raise an exception with a more
understandable error message.