Remove unneeded code from `SortedSet#eql?`
In the commit message for 952d11a357, I wrongly suggested that the code
for this method needed to be adjusted for Ruby 3.0. It seems that
rather, I had broken it when refactoring uses of `while true` loops to
use the `loop` method instead.

With that change, the final `rescue` clause became dead code.